requirements/base.txt
Django==1.10.6
Markdown==2.6.8
Pygments==2.2.0
pytz==2016.10
Pillow==4.1.0
django-allauth==0.32.0
django-contrib-comments==1.8.0
django-mptt==0.8.7
django-notifications-hq==1.2
django-haystack==2.6.1
django-model-utils==3.0.0
jieba==0.38
django-braces==1.11.0
django-imagekit==4.0.1
django-anymail==0.11.1
Whoosh==2.7.4